# docs/swagger.yaml
basePath: /
host: localhost:9001
info:
  contact:
    email: yono.tri@gmail.com
    name: API Support
  description: This is a documentation service for todo apps
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todos API
  version: "1.0"
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: create new todos with input payload.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: get todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: create new todos with input payload.
      parameters:
      - description: title todo
        in: query
        name: Title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create todos.
      tags:
      - todos
  /todos/{todoId}:
    delete:
      consumes:
      - application/json
      description: delete todo.
      parameters:
      - description: todo id
        in: path
        name: todoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete todo.
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: get single todo.
      parameters:
      - description: todo id
        in: path
        name: todoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: get single todos.
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: update new todos with input payload.
      parameters:
      - description: todo id
        in: path
        name: todoId
        required: true
        type: string
      - description: title todo
        in: query
        name: Title
        required: true
        type: string
      - description: 'todo done or yet '
        in: query
        name: Done
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update todos.
      tags:
      - todos
swagger: "2.0"